% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.wrap.loss.R
\name{lgb.wrap.loss}
\alias{lgb.wrap.loss}
\title{LightGBM loss function wrapper}
\usage{
lgb.wrap.loss(f)
}
\arguments{
\item{f}{Type: function. The function to wrap from LightGBM. Requires the following order of arguments for the function to work: \code{preds, labels}, and returns a vector of the same length of both the inputs.}
}
\value{
The wrapping function.
}
\description{
The wrapper works only if both the wrapper and the original loss metric are existing.
}
\examples{
library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)

cross_entropy <- function(preds, labels) {
  preds <- 1/(1 + exp(-preds))
  grad <- preds - labels
  hess <- preds * (1 - preds)
  return(list(grad = grad, hess = hess))
}
cross_entropy_wrap <- lgb.wrap.loss(f = cross_entropy)
params <- list(learning_rate = 1, min_data = 1, nthread = 1)

set.seed(1)
model <- lgb.cv(params,
                dtrain,
                2,
                nfold = 5,
                obj = cross_entropy_wrap,
                metric = "auc")
# [1]:	valid's auc:1+0 
# [2]:	valid's auc:1+0

}
