% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.f1.max.R
\name{metrics.f1.max}
\alias{metrics.f1.max}
\title{Maximum F1 Score (Precision with Sensitivity harmonic mean)}
\usage{
metrics.f1.max(preds, labels, threshold = FALSE)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{threshold}{Type: logical. Whether to return the threshold. Defaults to \code{FALSE}.}
}
\value{
A one or two element vector containing the maximum Maximum F1 Score (Precision with Sensitivity harmonic mean) for binary data, and the threshold used if requested. Returns \code{-1} in case of errors.
}
\description{
This function allows to use a custom thresholding method to maximize the Maximum F1 Score (Precision with Sensitivity harmonic mean).
}
