% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.fallout.max.R
\name{metrics.fallout.max}
\alias{metrics.fallout.max}
\title{Minimum Fall-Out (False Positive Rate)}
\usage{
metrics.fallout.max(preds, labels, threshold = FALSE)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{threshold}{Type: logical. Whether to return the threshold. Defaults to \code{FALSE}.}
}
\value{
A one or two element vector containing the minimum Fall-Out (False Positive Rate) for binary data, and the threshold used if requested. Returns \code{-1} in case of errors.
}
\description{
This function allows to use a custom thresholding method to minimize the Fall-Out (False Positive Rate).
}
