% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.logloss.R
\name{metrics.logloss}
\alias{metrics.logloss}
\title{Logarithmic Loss (logloss)}
\usage{
metrics.logloss(preds, labels, eps = 1e-15)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{eps}{Type: numeric. The shrinkage to apply between the 0 and 1's bounds.}
}
\value{
The Logarithmic loss.
}
\description{
This function computes the Logarithmic loss.
}
\examples{
set.seed(11111)
metrics.logloss(preds = runif(10000, 0, 1), labels = round(runif(10000, 1)))

}
