% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.logloss.unsafe.R
\name{metrics.logloss.unsafe}
\alias{metrics.logloss.unsafe}
\title{Logarithmic Loss (logloss)}
\usage{
metrics.logloss.unsafe(preds, labels, eps = 1e-15)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{eps}{Type: numeric. The shrinkage between the 0 and 1's bounds.}
}
\value{
The Logarithmic loss.
}
\description{
This function computes the Logarithmic loss without shrinking predictions (you may encounter infinite predictions).
}
\examples{
set.seed(11111)
metrics.logloss.unsafe(preds = runif(10000, 0, 1), labels = round(runif(10000, 1)))

}
