% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.precision.max.R
\name{metrics.precision.max}
\alias{metrics.precision.max}
\title{Maximum Precision (Positive Predictive Value)}
\usage{
metrics.precision.max(preds, labels, threshold = FALSE)
}
\arguments{
\item{preds}{Type: numeric. The predictions.}

\item{labels}{Type: numeric. The labels (0, 1).}

\item{threshold}{Type: logical. Whether to return the threshold. Defaults to \code{FALSE}.}
}
\value{
A one or two element vector containing the maximum Precision (Positive Predictive Value) for binary data, and the threshold used if requested. Returns \code{-1} in case of errors.
}
\description{
This function allows to use a custom thresholding method to maximize the Precision (Positive Predictive Value).
}
