% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.destroy.R
\name{parallel.destroy}
\alias{parallel.destroy}
\title{Parallel Destruction}
\usage{
parallel.destroy(cl = NULL, destroy = FALSE)
}
\arguments{
\item{cl}{Type: cluster or integer. A parallel cluster. When it is a number, creates a cluster with the specified number of parallel clusters. Defaults to \code{parallel::detectCores()}.}

\item{destroy}{Type: logical. Whether to attempt to destroy any cluster. Might have side effects. Defaults to \code{FALSE}.}
}
\value{
TRUE.
}
\description{
Attempt to kill the specified parallel cluster. Might have unexpected side effects when using the kill all command.
}
\examples{

\dontrun{
# Cannot pass CRAN checks. Disabled.
library(parallel)

# Generate 2 useless clusters of 8 threads
cl <- parallel::makeCluster(8)
cl <- parallel::makeCluster(8)

# Attempts to stop the second cluster
parallel.destroy(cl)

# Kills any clusters
parallel.destroy(destroy = TRUE)
}

}
