% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.read.R
\name{sparse.read}
\alias{sparse.read}
\title{SVMLight file reader}
\usage{
sparse.read(file, type = "CsparseMatrix", zero_based = TRUE, ncol = NULL)
}
\arguments{
\item{file}{Type: character. Path to your SVMLight file.}

\item{type}{Type: character. The matrix output type, which can be \code{CsparseMatrix}, \code{RsparseMatrix}, \code{TsparseMatrix}. Defaults to \code{CsparseMatrix}.}

\item{zero_based}{Type: logical. Whether columns are starting from \code{0} (\code{TRUE}), or from \code{1} (\code{FALSE}). Defaults to \code{TRUE}.}

\item{ncol}{Type: integer. The number of columns in the sparse matrix, useful in the case your data is very sparse and you already know the number of columns.}
}
\value{
The data from the SVMLight file.
}
\description{
Reads SVMLight sparse matrix file format.
}
\details{
Requires \code{Matrix} and \code{sparsio} packages.
}
\examples{
library(Matrix)
sparse_matrix <- sparseMatrix(i = 1:100, j = 1:100, x = 1:100)
temp_file <- tempfile(fileext = ".svm")
sparsio::write_svmlight(x = sparse_matrix, y = rep(1, ncol(sparse_matrix)), file = temp_file)
sparse_matrix_read <- sparse.read(temp_file)
identical(sparse_matrix, sparse_matrix_read$x)

}
