% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.write.R
\name{sparse.write}
\alias{sparse.write}
\title{SVMLight file writer}
\usage{
sparse.write(file, x, y = numeric(nrow(x)), zero_based = TRUE)
}
\arguments{
\item{file}{Type: character. Path to your SVMLight file.}

\item{x}{Type: any supported sparse matrix. The sparse matrix to write as SVMLight file format. Defaults to \code{numeric(nrow(x))}.}

\item{y}{Type: integer vector. The labels associated to the corresponding rows.}

\item{zero_based}{Type: logical. Whether columns are starting from \code{0} (\code{TRUE}), or from \code{1} (\code{FALSE}). Defaults to \code{TRUE}.}
}
\value{
TRUE when the function returns.
}
\description{
Writes SVMLight sparse matrix file format.
}
\details{
Requires \code{Matrix} and \code{sparsio} packages.
}
\examples{
library(Matrix)
sparse_matrix <- sparseMatrix(i = 1:100, j = 1:100, x = 1:100)
temp_file <- tempfile(fileext = ".svm")
sparse.write(file = temp_file, x = sparse_matrix, y = rep(1, ncol(sparse_matrix)))
sparse_matrix_read <- sparsio::read_svmlight(temp_file)
identical(sparse_matrix, sparse_matrix_read$x)

}
