% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lgb.wrap.metric.R
\name{lgb.wrap.metric}
\alias{lgb.wrap.metric}
\title{LightGBM evaluation metric wrapper}
\usage{
lgb.wrap.metric(f, name, higher_better = FALSE)
}
\arguments{
\item{f}{Type: function. The function to wrap from LightGBM. Requires the following order of arguments for the function to work: \code{preds, labels}, and returns a single value.}

\item{name}{Type: character. The evaluation metric name which is printed and used for LightGBM variables.}

\item{higher_better}{Type: logical. Whether a higher value means it is better. Defaults to \code{FALSE}.}
}
\value{
The wrapping function.
}
\description{
The wrapper works only if both the wrapper and the original evaluation metric are existing.
}
\examples{
library(lightgbm)
data(agaricus.train, package = "lightgbm")
train <- agaricus.train
dtrain <- lgb.Dataset(train$data, label = train$label)

evalerror <- function(preds, labels) {
  return(mean(abs(preds - labels)))
}
evalerror_wrap <- lgb.wrap.metric(f = evalerror, "mae", FALSE)
params <- list(learning_rate = 1, min_data = 1, nthread = 1)

set.seed(1)
model <- lgb.cv(params,
                dtrain,
                2,
                nfold = 5,
                objective = "regression",
                eval = evalerror_wrap)
# [1]:	valid's mae:1.19113e-08+3.07441e-09 
# [2]:	valid's mae:2.18733e-16+1.80747e-16

}
